Keep validating files against schema, even after errors
It might be better to keep checking files even after the first invalid
file so you can get a complete list of problem files instead of running
`make test-schemas`, fixing the file it complains about, running make
again, and so forth. So keep track of number of invalid files and only
crap out at the end.