Change get_json_files to support subfolders as root path
Before it was not possible to provide a subfolder to scripts
like `.tests/render_rest.py`. Maybe it was never intended,
but I find it helpful for testing only the conf folder, that
I am currently working on.

With this change it is possible to only run tests on the json
file of a particular conference folder, e.g.
`python .tests/render_rest.py -d ./python-web-conf-2020`
The change does not interfere with the original functionality
of the script.