Exclude schema definition files from schema validation checks (#1343)
Currently, the category schema definition file is included in the list
of files that are validated against it. This causes validation errors
when the schema validation's strictness is increased, e.g. by
disallowing additional properties in #1172. This change excludes schema
JSON files from the schema test script to resolve future validation
errors.