Allow start_date, end_date, and URL in category schema (#1344)
Add start and end date fields as well as URL to the category schema.
These properties are common across many categories supported here and
already present in a handful of data files.

This change also populates start and end date information based on the
recording dates in video files for a couple of categories that have
these fields present but do not pass the JSON Schema date format
validation.

References #268 and #1172.