Remove speaker names from beginnings of talk descriptions
Some video descriptions here (and on the original YouTube
descriptions) include speaker names before the main body.
These speaker names are not present on the schedule pages
for these presentations, and PyVideo stores speaker names
in a dedicated field, so this commit removes the prefixed
names in favor of matching the original website and using
PyVideo's data model.