Validate directory layout with pathschema
This is a first pass at validating the PyVideo data directory layout
with a schema validation tool.

The included schema validates the data project as a whole and could
be simplified by restructuring the repository to move data into a
non-root directory to avoid the need to validate the existence of any
other directories. This would be a good next step and require a
coordinated change with the pyvideo/pyvideo repository.

Resolves #1125.